prettier

+++ packages/onchain/contracts/test/MockUniRouter.sol
@@ -1,23 +1,21 @@
 pragma solidity 0.6.6;
 
 contract MockUniRouter {
-
   address uniFactory;
 
-  function factory() external view returns (address){
+  function factory() external view returns (address) {
     return uniFactory;
   }
 
   function swapExactTokensForTokens(
-    uint amountIn,
-    uint amountOutMin,
+    uint256 amountIn,
+    uint256 amountOutMin,
     address[] calldata path,
     address to,
-    uint deadline
-  ) external returns (uint[] memory amounts){
-    amounts = new uint[](2);
+    uint256 deadline
+  ) external returns (uint256[] memory amounts) {
+    amounts = new uint256[](2);
     amounts[0] = amountIn;
     amounts[1] = amountOutMin;
   }
-
 }
